Adding Inference File for Gesture Classifier
