+++ training (initial)/collect_from_camera.py
@@ -1,7 +1,7 @@
 import cv2
 import numpy as np
 import os
-from models.gesture_classifier import GestureClassifier
+from test.models.gesture_classifier_2 import GestureClassifier
 
 def setup_data_directories(class_names):
     """
